backend/requirements.txt
langgraph==0.2.28
langchain-openai==0.2.1
langchain-core==0.3.7
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
psutil==5.9.8

